Fixed two defects & segmented the code base into two versions.
----Version 1: ISS_FIT_Samba_Test_Data
Includes the bug fixes, no other changes to the code. App will sync normally with Samba Server for test data.

----Version 2: ISS_FIT_Pre-load_Test_Data
Includes the bug fixes, but this version has the test data loaded into the iOS app as the app is compiled and built, vs Version 1 which pulls the test data from the Samba Server. Also, the friendly error when you load the app and no Samba Server is available "We are unable to sync this iPad with the central..." should no longer appear (this version will be for testing the app without a Samba server).